// client/src/app/entities/sitter.ts
// export class Sitter {
//   id: string;
//   name: string;
//   description: string;
//   quote: string;
//   minimumPrice: number;
//   maximumPrice: number;
//   images: string[];
//   recentOnlineDate: Date;
//   strengths: string[];
//   availability: string;
//   reviews: number;
//   verified: number;
//   location: string;
// }

export class Sitter {
  id: string;
  data: data;
}

export class data {
  age: number;
  name: string;
  description: string;
  quote: string;
  minimumPrice: number;
  maximumPrice: number;
  images: string[];
  recentOnlineDate: Date;
  strengths: string[];
  availability: string;
  reviews: number;
  verified: number;
  location: string;
}